Enemy Inheritance, Bullet Damage, and Organization
Updated BaseEnemy and RatController scripts.  BaseEnemy now handles all
underlying animation/collision/health/death logic for enemies while
RatController handles behavior specific to the rat enemy.  Controller
scripts for future enemies will only need to define the behavior of that
enemy (movement, projectiles, etc.).  Note that future enemy scripts
should include a call to the base start function in Start() and check
the IsDead property in the beginning of Update(), just like in the
RatController.  The death animation transition for future enemies should
also be based on an "isDead" boolean parameter, just like the rat's is.
I also modified the BulletController script to include a "damagePoints"
variable to store the hit damage information for the different
projectile prefabs we'll have.  The axe currently does 1 damage, and a
made a beer bullet for testing that does 2 damage (one hit kill for the
rats).  This damagePoints value is used in the BaseEnemy controller to
deduct the correct amount of health from an enemy when it is hit by a
certain projectile.
Lastly, I also organized the prefab folder into three sections for the
three types of prefabs we currently have.